переработана структура проекта
