+++ lib/data/reader_json.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/services.dart';
+
+Future<String> readJsonFile(String filePath)  {
+  return rootBundle.loadString(filePath);
+}
