+++ lib/listeners/map_object_tap_listener.dart
@@ -1,12 +1,14 @@
 import 'package:yandex_maps_mapkit/mapkit.dart';
 
 final class MapObjectTapListenerImpl implements MapObjectTapListener {
-  final bool Function(MapObject, Point) onMapObjectTapped;
+  // final bool Function(MapObject, Point) onMapObjectTapped;
 
-  const MapObjectTapListenerImpl({required this.onMapObjectTapped});
+  // const MapObjectTapListenerImpl({required this.onMapObjectTapped});
 
   @override
   bool onMapObjectTap(MapObject mapObject, Point point) {
-    return onMapObjectTapped(mapObject, point);
+    print("Tapped the placemark: Point(latitude: ${point.latitude}, longitude: ${point.longitude})");
+    return true;
+    // return onMapObjectTapped(mapObject, point);
   }
 }
