+++ lib/views/map_page.dart
@@ -1,0 +1,60 @@
+import 'package:flutter/material.dart';
+
+// import '../listeners/map_object_tap_listener.dart';
+import 'package:my_mystery_city/controllers/map_state.dart';
+import 'package:my_mystery_city/data/reader_json.dart';
+
+import 'package:yandex_maps_mapkit/mapkit.dart';
+import 'package:yandex_maps_mapkit/mapkit_factory.dart';
+import 'package:yandex_maps_mapkit/yandex_map.dart';
+
+
+MapWindow? mapWindow_;
+
+
+class MapPage extends StatefulWidget {
+  const MapPage({super.key});
+  @override
+  State<MapPage> createState() => _MapPageState();
+}
+
+class _MapPageState extends State<MapPage> {
+  var defaultPoint = Point(latitude: 56.837716, longitude: 60.596828); // убрать с костылём зума на екб
+
+  @override
+  Widget build(BuildContext context) {
+  return MaterialApp(
+    theme: Theme.of(context),
+    home: Scaffold(
+      body: Stack(
+        children: [
+          YandexMap(
+            onMapCreated: (mapWindow) async {
+              mapWindow_ = mapWindow;
+              mapWindow.map.setMapStyle(await readJsonFile("assets/style/style_map.json"));
+              await makePoints(mapWindow_);
+              // Костыль, чтобы пока позиция пользователя загружалась, карта заранее смотрела на Екатеринбург, а не на весь мир
+              mapWindow_!.map.move(CameraPosition(defaultPoint, zoom: 12.5, azimuth: 0.0, tilt: 30.0));
+              await moveToUserLocation(mapWindow_);
+              mapkit.onStart();
+            },
+          ),
+          Positioned(
+            bottom: 50,
+            right: 0,
+            child: ElevatedButton(
+              onPressed: () {
+                moveToUserLocation(mapWindow_);
+              },
+              style: ElevatedButton.styleFrom(
+                shape: CircleBorder(),
+              ),
+              child: Icon(Icons.near_me),
+            ),
+          ),
+        ],
+      ),
+    ),
+  );
+}
+}
